simulations/requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.2
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
setuptools==78.1.1
six==1.17.0
wheel==0.45.1
pandas==2.2.3
pickleshare==0.7.5
